itoa-memalloc-putnbr-strlen-strmap-strmapi-.h

+++ LIBFT/ft_strlen.c
@@ -13,9 +13,9 @@
 #include <stdio.h>
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
